Make sure university degree status has precedence
On TDA current live site (soon to be the old parameter) the
university degree status param is used in conjuction with
the degree required show all courses.

Now we are moving TDA to the degree required we need to make
sure TDA has precedence when copying the old url to the new results
page.

Example:

Visit
https://find-review-4869.test.teacherservices.cloud/v2/results?university_degree_status=false
expects the "No degree required" option to be checked

Visit
https://find-review-4869.test.teacherservices.cloud/v2/results?university_degree_status=false&degree_required=show_all_courses
also expects the "No degree required" option to be checked