Add location and subject search feature
The subject field will use subject_code behing the curtains
and location will be used to get the coordinates from google (and
cached!).

Also I had to add a specific logic to subject autocomplete to fix this
issue:

1. Type "Pri" on subject
2. Chooses "Primary"
3. Click "Search"
4. Removes "Primary"
5. Click "Search"

That's why we need to clean the select element after is empty.