Add provider search to the results page form
This commit adds the ability to search by provider name to the form
in the results page.

The field includes an autocomplete that drops down with all the list of providers

Content
<details> Search by training provider

<label> Enter provider name

Initially I did via a remote call via JS, but it sounds too slow
for a simple search, so I optin to have the list available at the
client but cache to not hit the database again for a hour.

The one hour expiration was arbitrary and I need to align with tech
lead if this is a good expiration rule (or should be higher or lower).