Add course start date filter
In the last round of pre-filtering testing, we tested a
new ‘Start date' filter, which tested well.

In 2024, 91% of start dates fell in September, 5% in October, 2% in July,
and very few courses scattered elsewhere.
Regarding % applications sent, September courses are proportionally
more popular, having 94% of applications.

we decided that although this feature targets a small portion of users,
it is incredibly valuable to them, and aligns with the Flexible ITT strategy.