Increase the number of pods
Next week we are releasing pre-filtering work and we are slightly increasing the number of pods.

The new /results page on Find needs to hit Google API for the location searches on the first time only, then it caches for 30 days. This is IO bound but with 5 concurrency as limitation, a low risk but better to increase slightly for the first week and decrease if needed after.

Increasing the number of pods slightly just as manner of cautious while we monitor the DB CPU and application performance next week.