Show only open courses by default on Find searches
**Problem**

When performing any search and landing on the search results page,
the filter 'Only show courses open for applications' is unchecked by default.

This means that candidates will see courses that are closed, which may
result in a frustrating experience. There may be only certain scenarios
where candidates may want to do this
(we are considering whether we need this filter at all),
but I think we can be confident in saying the majority of candidates
want to see course that they can apply to.