Update course ratifying provider description
Use the `accrediting_provider` instance directly instead of trying to re-lookup the provider.

The error we were seeing is that courses from previous recruitment cycles may have providers that are not in the current cycle.

We can safely return the description of a ratifying provider's details from a previous recruitment cycle.