Remove can_edit_current_and_next_cycles
This can_edit_current_and_next_cycles: false is when
there is no next recruitment cycle with available from publish date
set, so no record automatically gives us that.

For the ones that needs, we need to create the recruitment cycle
record and set the proper "available in publish from" date.