Fix prototype deployments
Unfortunately, GitHub doesn't natively provide a method of
conditionally triggering workflows based on the source branch.

Given a workflow trigger of:

```yaml
on:
  pull_request:
    branches:
      - prototype/**
```

This will run the workflow on pull requests _into_ the
`prototype/**` branches, not _from_.

Therefore, we need to run this workflow on all branches and skip
the jobs below based on conditions tied to the source branch.