Create "Added schools" from register page
We are making changes to Publish this cycle, removing non-validated
schools and adding schools from Register to providers accounts.

We need to clearly communicate these changes to providers in the service.

This commit creates a new page view in Publish that lists the schools
added to the providers account from Register.

The back link will take them back to the Schools landing page.