Ignore some providers during register school import
As part of the work to add all placement schools from Register to
Publish we have identified provider who will be receiving
many new schools.
Top of the list.

The provider run both regular ITT, as well as
High Potential ITT (HPITT) which is recruited off-service,
we should not be adding HPITT only schools to their Publish account.