Change service unavailable page content
Find and Publish will be down between 8am and 9am on 16 October, so we will need to have a ‘Service is unavailable’ page during that time.

There will be just one page for both Find and Publish.
This content will also be used on Apply and Manage, which will have downtime at the same time.